deleted:    Nurumov_r/task02/src/task02.py 	new file:   Nurumov_r/task02/src/task02/__init__.py 	new file:   Nurumov_r/task02/src/task02/tree.py 	new file:   Nurumov_r/task02/test/test_tree.py
